modules/location-kafka-service/requirements.txt
GeoAlchemy2==0.12.5
greenlet==2.0.1
grpcio==1.51.1
grpcio-tools==1.51.1
kafka-python==2.0.2
packaging==22.0
protobuf==4.21.11
psycopg2-binary==2.9.5
SQLAlchemy==1.4.45